+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.18.0] - 2023-07-11
+### Added
+- Endpoint `api/load/file` for uploading static files
+
+### Changed
+- SvgLoadHandler to FileLoadHandler in order to reuse the same class for both endpoints (svg,static)
+
 ## [1.17.3] - 2023-03-30
 ### Changed
 - Endpoint `api/group/dashs` will not return dashboards' body since this release.